fixed refresh issue

+++ LunarCard.h
@@ -10,7 +10,7 @@ class LunarCard {
   public:
     String cardname;
     int index;
-    unsigned int msLoaded=0;
+    unsigned long msLoaded=0;
     vector<Transition *> transitions;
     CardType type;
     String toString();
